+++ CHANGES.md
@@ -4,10 +4,13 @@ moodle-theme_boost_union
 Changes
 -------
 
-### v4.5-r11
+### Unreleased
 
 * 2025-03-27 - Improvement: Add admin main navigation to smart menu items page as well, resolves #882.
 * 2025-03-27 - Improvement: Add tertiary navigation to Boost Union admin settings pages to allow jumping from one settings page to another, resolves #876.
+
+### v4.5-r11
+
 * 2025-03-26 - Bugfix: Footer button got pushed upwards even though no bottom menu bar was present, resolves #784.
 * 2025-03-25 - Bugfix: If a course category is hidden, the "go to course" button (in the course cards view) got a wrong color, resolves #877.
 * 2025-03-25 - Bugfix: Make sure that the sticky category headers (in the course cards view) does not cover the more menu, resolves #866.